+++ api/bff.py
@@ -26,7 +26,6 @@ class Game(BaseModel):
 class Guess(BaseModel):
     user_id: str
     guess: str
-    game_id: Optional[int] = int(datetime.date.today().strftime("%Y%m%d"))
 
 
 # get user_id from stats service
@@ -56,14 +55,30 @@ def create_game(game: Game):
         print(response.text)
 
 
+def getGame(game: Game):
+    game_id = game.get("game_id")
+    user_id = game.get("user_id")
+    headers = {"content-type": "application/json"}
+    with httpx.Client(base_url="http://localhost:9999/api/state") as client:
+        response = client.get(f"users/{user_id}/game/{game_id}", headers=headers)
+        print(response.status_code)
+        text = json.loads(response.text)
+        print(json.dumps(text, indent=4))
+
+
 # TODO: check if user has guesses remaining in this game
-async def isValidGame():
-    async with httpx.AsyncClient(
-        app=game_state_service, base_url="http://localhost:9999/api/state"
-    ) as client:
-        r = await client.get("/game/")
-        print(r.status_code)
-        print(json.dumps(json.loads(r.text), indent=4))
+def isValidGame(game: Game):
+    game_id = game.get("game_id")
+    user_id = game.get("user_id")
+    headers = {"content-type": "application/json"}
+    with httpx.Client(base_url="http://localhost:9999/api/state") as client:
+        response = client.get(f"users/{user_id}/game/{game_id}", headers=headers)
+        text = json.loads(response.text)
+        status = text.get("status")
+        if status == "finished":
+            return False
+        else:
+            return True
 
 
 # TODO: check if word is valid guess
@@ -86,14 +101,18 @@ async def check_guess():
         print(json.dumps(json.loads(r.text), indent=4))
 
 
-# TODO: post user's guess to game
-async def update_game():
-    async with httpx.AsyncClient(
-        app=game_state_service, base_url="http://localhost:9999/api/state"
-    ) as client:
-        r = await client.post("/game/")
-        print(r.status_code)
-        print(json.dumps(json.loads(r.text), indent=4))
+# post user's guess to game
+def add_guess(game: Game, guess: str):
+    # load game data
+    game_id = game.get("game_id")
+    user = game.get("user_id")
+    # build Guess model to pass to game_state service
+    current_guess = Guess(user_id=user, guess=guess).json()
+    headers = {"content-type": "application/json"}
+    with httpx.Client(base_url="http://localhost:9999/api/state") as client:
+        response = client.post(f"/game/{game_id}", data=current_guess, headers=headers)
+        print(response.status_code)
+        print(json.dumps(json.loads(response.text), indent=4))
 
 
 # TODO: update user stats if this game is finished
@@ -123,16 +142,27 @@ async def getStats():
 #   - game doesn't exist
 #   - game has ended
 #   - word isn't valid
-def new_guess(guess=Guess):
-    user_id = guess.get("user_id")
-    game_id = guess.get("game_id")
-    word = guess.get("guess")
+def new_guess(game_id: int, guess=Guess):
+    # get user from guess request body
+    user = guess.get("user_id")
+    # create game object
+    game = Game(user_id=user, game_id=game_id).json()
+    # get guess string from guess request body
+    current_guess = guess.get("guess")
 
     # check if game is valid
-    isValidGame(user_id, game_id)
+    print("\nisValidGame():")
+    if isValidGame(json.loads(game)):
+        print("True")
+    else:
+        print("False")
 
     # check if word is valid
-    isValidWord()
+    # isValidWord()
+
+    # add new guess to game
+    print("\nadd_guess():")
+    add_guess(json.loads(game), current_guess)
 
 
 # TODO: workflow for adding a new game
@@ -140,24 +170,32 @@ def new_guess(guess=Guess):
 # output: game state of this game
 # error conditions:
 #   - game id already exists (conflict)
-
-
 def new_game(username: User):
+    # game_id test value - when not using default (today)
+    test_gameid = 20220517
+    # get user_id
     user = getUser(username)
-    # print for debugging
+    # print values for getUser response
     print("getUser():\n\t", user, "\n")
     user = user.get("user_id")
     # create new game object using user and default game_id
     game = Game(user_id=user).json()
+    game = Game(user_id=user, game_id=test_gameid).json()
     # print for debugging
     print("Game model object:\n\t", game, "\n")
+    print("\ncreate_game():")
     # pass new game object and create new game
     create_game(game)
+    print("\ngetGame():")
+    getGame(json.loads(game))
+    guess = Guess(user_id=user, guess="helps").json()
+    new_guess(test_gameid, json.loads(guess))
 
 
 # test new game() for endpoint
 user1 = "christina22"
 new_game(user1)
+
 
 # TEST async
 # asyncio.run(getUser(user1))